Mass update of EX drivers. Bias now is completely handled as a float.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@9803 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/ex/ST/l3gd20.c
@@ -169,7 +169,7 @@ static msg_t sample_bias(void *ip) {
   return msg;
 }
 
-static msg_t set_bias(void *ip, int32_t *bp) {
+static msg_t set_bias(void *ip, float *bp) {
   uint32_t i;
   
   osalDbgCheck((ip != NULL) && (bp !=NULL));
